Rename label_maps module. Snake case is used consistently throughout the crate.
